refactor(profile): use modularized code
